+++ app/src/main/res/values/strings.xml
@@ -444,6 +444,7 @@
     <string name="device_settings_aap_unavailable_description">This phone\'s Bluetooth does not support the connection required for advanced AirPods settings.</string>
     <string name="device_settings_category_sound_label">Sound</string>
     <string name="device_settings_category_controls_label">Controls</string>
+    <string name="device_settings_category_smart_features_label">Smart features</string>
     <string name="device_settings_nc_one_airpod_label">Noise Cancellation with One AirPod</string>
     <string name="device_settings_nc_one_airpod_description">Keep noise cancellation active with only one AirPod</string>
     <string name="device_settings_personalized_volume_label">Personalized Volume</string>
